Moved testfactories to fake package
